Refactor interface DTO's and rut id to long

+++ TallerJava/src/main/java/org/tallerJava/purchaseModule/domain/PurchaseCommerce.java
@@ -17,7 +17,7 @@ import java.util.List;
 @Table(name = "purchase_Commerce")
 public class PurchaseCommerce {
     @Id
-    private int rut;
+    private long rut;
 
     @OneToMany(mappedBy = "commerce", cascade = CascadeType.ALL)
     private List<Purchase> purchases = new ArrayList<>();
@@ -26,7 +26,7 @@ public class PurchaseCommerce {
 
     private double totalSalesAmount = 0d;
 
-    public PurchaseCommerce(int rut) {
+    public PurchaseCommerce(long rut) {
         this.rut = rut;
     }
